Example INCOMMON dataset.

Genomic data from sample P-0002081-T01-IM3 of the MSK-MET cohort.

Usage

data(example_data)

Format

A named list with the following components:

mutations

A data frame containing mutations.

sample

A string indicating the name of the sample.

tumor_type

A string indicating the tumor type.

purity

A numeric indicating the fraction of tumor in sample.

The columns in the data frame (mutations) are:

chr

Chromosome harbouring the mutation.

from

Genomic start position of the mutation.

to

Genomic end position of the mutation.

ref

Reference allele.

alt

Alternative allele.

gene

Gene harbouring the mutation.

NV

Number of reads with the variant

DP

Total number of reads (sequencing depth)

VAF

Variant allele frequency of the mutation.

Source

MSK-MET at cBioPortal https://www.cbioportal.org/study/summary?id=msk_met_2021

Details

This list consists the following components:

mutations: A data frame containing mutations from sample P-0002081-T01-IM3 of the MSK-MET cohort with columns chr, from, to, ref, alt, gene, NV, DP and VAF.

sample: The name of the sample.

tumor_type: The tumor type of the sample.

purity: The purity of the tumor sample (fraction).

Examples

data(example_data)
example_data
#> $data
#> # A tibble: 4 × 9
#>   chr       from       to ref   alt   gene       NV    DP   VAF
#>   <chr>    <dbl>    <dbl> <chr> <chr> <chr>   <int> <int> <dbl>
#> 1 chr12 25398285 25398285 C     A     KRAS      378   743 0.509
#> 2 chr17  7577139  7577139 G     A     TP53      116   246 0.472
#> 3 chr19  1221293  1221293 C     A     STK11     122   260 0.469
#> 4 chr19 11141472 11141473 -     C     SMARCA4   133   271 0.491
#> 
#> $sample
#> [1] "P-0002081-T01-IM3"
#> 
#> $tumor_type
#> [1] "LUAD"
#> 
#> $purity
#> [1] 0.6
#> 
if (FALSE) {
# Create an example data
example <- list(
mutations <- data.frame(
 chr = c("chr1", "chr14", "chr17"),  # Chromosome column
 from = c(16265908, 105246551, 7578503),  # Start position column
 to = c(16265908, 105246551, 7578518),  # End position column
 ref = c("A", "C", "CAGGGCAGGTCTTGGC"),  # Reference allele column
 alt = c("T", "T", "-"),  # Alternate allele column
 gene = c("SPEN", "AKT1", "TP53")  # Gene column
),
  sample = "P-0000004-T01-IM3",
  purity = 0.5
)
}