This function returns, for each mutation, an interpreted state depending on the mutant gene role and its copy-number and multiplicity, among Mutant TSG with or without LOH, and Mutant oncogene with or without amplification. In addition, it assigns an interpreted genome to each sample by integrating the interpreted state of each mutant gene in the sample.

This function returns, for each mutation, an interpreted state depending on the mutant gene role and its copy-number and multiplicity, among Mutant TSG with or without LOH, and Mutant oncogene with or without amplification. In addition, it assigns an interpreted genome to each sample by integrating the interpreted state of each mutant gene in the sample.

Usage

genome_interpreter(x)

Arguments

x

An object of class 'INCOMMON' containing the classification results, as produced by function classify.

Value

An object or a list of class INCOMMON with additional columns in classification.

Examples

# First load example classified data
data(MSK_classified)
# Note the outputs to screen
genome_interpreter(MSK_classified)
#> ℹ There are 20289 different genotypes
#> ℹ The most abundant genotypes are:
#> • Mutant TP53 with LOH (562 Samples, Frequency 0.02)
#> • Mutant KRAS without AMP (199 Samples, Frequency 0.01)
#> • Mutant KRAS without AMP,Mutant TP53 with LOH (149 Samples, Frequency 0.01)
#> ── [ INCOMMON ]  176075 PASS mutations across 24117 samples, with 290 mutant gen
#> ℹ Average sample purity: 0.4
#> ℹ Average sequencing depth: 648
#> ── [ INCOMMON ]  Classified mutations with overdispersion parameter 0.01 and ent
#> ℹ There are: 
#> • N = 63649 mutations (HMD)
#> • N = 11905 mutations (LOH)
#> • N = 35087 mutations (CNLOH)
#> • N = 22803 mutations (AM)
#> • N = 42629 mutations (Tier-2)
#> # A tibble: 176,073 × 21
#>    sample tumor_type purity genotype chr     from     to ref   alt      DP    NV
#>    <chr>  <chr>       <dbl> <chr>    <chr>  <dbl>  <dbl> <chr> <chr> <int> <int>
#>  1 P-000… BRCA          0.5 SPEN Ti… chr1  1.63e7 1.63e7 A     T       473    73
#>  2 P-000… BRCA          0.5 SPEN Ti… chr14 1.05e8 1.05e8 C     T       446   244
#>  3 P-000… BRCA          0.5 SPEN Ti… chr17 7.58e6 7.58e6 CAGG… -       267    58
#>  4 P-000… BRCA          0.4 Mutant … chr2  2.95e7 2.95e7 C     T       243    47
#>  5 P-000… BRCA          0.4 Mutant … chr3  1.79e8 1.79e8 G     A       136    56
#>  6 P-000… BRCA          0.4 Mutant … chr6  1.52e8 1.52e8 T     A       123    45
#>  7 P-000… BRCA          0.4 Mutant … chr12 5.81e7 5.81e7 C     G       430    84
#>  8 P-000… BRCA          0.4 Mutant … chr17 7.58e6 7.58e6 G     A       192    61
#>  9 P-000… BRCA          0.4 Mutant … chr17 5.64e7 5.64e7 C     G       854    79
#> 10 P-000… BRCA          0.4 Mutant … chr10 8.11e6 8.11e6 -     G       453    75
#> # ℹ 176,063 more rows
#> # ℹ 10 more variables: VAF <dbl>, gene <chr>, HGVSp_Short <chr>,
#> #   gene_role <chr>, id <chr>, label <chr>, state <chr>, posterior <dbl>,
#> #   entropy <dbl>, class <chr>